app/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
docopt==0.6.2
google-auth==2.8.0
google-auth-oauthlib==0.5.2
greenlet==1.1.2
gspread==5.4.0
html-tag-names==0.1.2
html-void-elements==0.1.0
html2text==2020.1.16
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.4
lxml==4.9.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.0
pandas==1.4.2
pathspec==0.9.0
psycopg==3.0.15
psycopg2-binary==2.9.3
putz==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pypi-search==1.2.1
pyTelegramBotAPI==4.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.9
yapf==0.32.0
zipp==3.8.0
